requirements.txt
Django>=4.2.0
newrelic>=11.0.0
locust>=2.34.0
beautifulsoup4>=4.12.0
psutil>=5.9.0
python-dotenv>=1.0.0

